# LctMonolith/openapi-gamification.yaml
openapi: 3.0.3
info:
  title: LctMonolith Gamification API
  version: 1.0.0
  description: |
    Comprehensive REST API for gamification module (players, missions, ranks, skills, rewards, dialogue, inventory, store, analytics, notifications, auth).
    Authentication via JWT Bearer (Authorization: Bearer <token>). Admin endpoints require role=Admin.
servers:
  - url: https://localhost:5001
    description: Local HTTPS
  - url: http://localhost:5000
    description: Local HTTP
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Id: { type: string, format: uuid }
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresAt: { type: string, format: date-time }
        refreshExpiresAt: { type: string, format: date-time, nullable: true }
...
    SkillProgress:
      type: object
      properties:
        skillId: { $ref: '#/components/schemas/Id' }
        skillTitle: { type: string }
        previousLevel: { type: integer }
        newLevel: { type: integer }
    MissionCompletionResult:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        experienceGained: { type: integer }
        manaGained: { type: integer }
        skillsProgress: { type: array, items: { $ref: '#/components/schemas/SkillProgress' } }
        unlockedMissions: { type: array, items: { $ref: '#/components/schemas/Id' } }
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
                firstName: { type: string }
                lastName: { type: string }
              required: [email,password]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email,password]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
              required: [refreshToken]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
  /api/auth/revoke:
    post:
      tags: [Auth]
      security: [ { bearerAuth: [] } ]
      summary: Revoke refresh token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '204': { description: No Content }
  /api/auth/me:
    get:
      tags: [Auth]
      security: [ { bearerAuth: [] } ]
      summary: Current user id
      responses:
        '200': { description: OK }
...
  /api/mission-categories:
    get:
      tags: [MissionCategories]
      security: [ { bearerAuth: [] } ]
      summary: List mission categories
      responses: { '200': { description: OK } }
    post:
      tags: [MissionCategories]
      security: [ { bearerAuth: [] } ]
      summary: Create mission category (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
              required: [title]
      responses: { '201': { description: Created } }
      x-roles: [Admin]
  /api/mission-categories/{id}:
    get:
      tags: [MissionCategories]
      security: [ { bearerAuth: [] } ]
      summary: Get mission category
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK }, '404': { description: Not Found } }
    put:
      tags: [MissionCategories]
      security: [ { bearerAuth: [] } ]
      summary: Update mission category (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
              required: [title]
      responses: { '200': { description: OK }, '404': { description: Not Found } }
      x-roles: [Admin]
    delete:
      tags: [MissionCategories]
      security: [ { bearerAuth: [] } ]
      summary: Delete mission category (Admin)
      responses: { '204': { description: No Content }, '404': { description: Not Found } }
      x-roles: [Admin]
...
  /api/missions/category/{categoryId}:
    get:
      tags: [Missions]
      security: [ { bearerAuth: [] } ]
      summary: Missions by category
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/missions/player/{playerId}/available:
    get:
      tags: [Missions]
      security: [ { bearerAuth: [] } ]
      summary: Available missions for player
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/missions/{missionId}/rank-rules:
    get:
      tags: [Missions]
      security: [ { bearerAuth: [] } ]
      summary: Rank rules for mission
      parameters:
        - in: path
          name: missionId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/missions:
    post:
      tags: [Missions]
      security: [ { bearerAuth: [] } ]
      summary: Create mission (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                missionCategoryId: { $ref: '#/components/schemas/Id' }
                parentMissionId: { $ref: '#/components/schemas/Id', nullable: true }
                expReward: { type: integer }
                manaReward: { type: integer }
              required: [title,missionCategoryId,expReward,manaReward]
      responses: { '201': { description: Created } }
      x-roles: [Admin]
  /api/missions/{id}:
    put:
      tags: [Missions]
      security: [ { bearerAuth: [] } ]
      summary: Update mission (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                missionCategoryId: { $ref: '#/components/schemas/Id' }
                parentMissionId: { $ref: '#/components/schemas/Id', nullable: true }
                expReward: { type: integer }
                manaReward: { type: integer }
              required: [title,missionCategoryId,expReward,manaReward]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK }, '404': { description: Not Found } }
      x-roles: [Admin]
    delete:
      tags: [Missions]
      security: [ { bearerAuth: [] } ]
      summary: Delete mission (Admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '204': { description: No Content }, '404': { description: Not Found } }
      x-roles: [Admin]
  /api/missions/{missionId}/complete:
    post:
      tags: [Missions]
      security: [ { bearerAuth: [] } ]
      summary: Complete mission
      parameters:
        - in: path
          name: missionId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId: { $ref: '#/components/schemas/Id' }
                proof: { nullable: true }
              required: [playerId]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MissionCompletionResult' } } } }, '400': { description: Bad Request } }
...
  /api/dialogue/mission/{missionId}:
    get:
      tags: [Dialogue]
      security: [ { bearerAuth: [] } ]
      summary: Get dialogue by mission
      parameters:
        - in: path
          name: missionId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK }, '404': { description: Not Found } }
  /api/dialogue/message/{messageId}:
    get:
      tags: [Dialogue]
      security: [ { bearerAuth: [] } ]
      summary: Get dialogue message
      parameters:
        - in: path
          name: messageId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK }, '404': { description: Not Found } }
  /api/dialogue/message/{messageId}/options:
    get:
      tags: [Dialogue]
      security: [ { bearerAuth: [] } ]
      summary: Get dialogue response options
      parameters:
        - in: path
          name: messageId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/dialogue/message/{messageId}/respond:
    post:
      tags: [Dialogue]
      security: [ { bearerAuth: [] } ]
      summary: Respond to dialogue
      parameters:
        - in: path
          name: messageId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                responseOptionId: { $ref: '#/components/schemas/Id' }
                playerId: { $ref: '#/components/schemas/Id' }
              required: [responseOptionId,playerId]
      responses: { '200': { description: OK } }
  /api/dialogue:
    post:
      tags: [Dialogue]
      security: [ { bearerAuth: [] } ]
      summary: Create dialogue (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                missionId: { $ref: '#/components/schemas/Id' }
                initialDialogueMessageId: { $ref: '#/components/schemas/Id' }
                interimDialogueMessageId: { $ref: '#/components/schemas/Id' }
                endDialogueMessageId: { $ref: '#/components/schemas/Id' }
              required: [missionId,initialDialogueMessageId,interimDialogueMessageId,endDialogueMessageId]
      responses: { '201': { description: Created } }
      x-roles: [Admin]
...